==> 93.Revision/01.StandardAAMix/04.Signal_Selecting_and_Predicting.Rmd <==
---
title: "Figure 5 c"
author: "Chao Tang"
date: 'Report created: `r Sys.Date()`'
output: 
  html_document: 
    code_folding: "hide"
    toc: true
    toc_depth: 4
    toc_float: 
      collapsed: false
    number_sections: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, 
                      warning = FALSE, 
                      message = FALSE, 
                      fig.align = 'center')
knitr::opts_knit$set(root.dir = "/mnt/raid61/Personal_data/tangchao/AANanopore")
```

```{r warning=FALSE}
library(data.table)
library(Biostrings)
library(patchwork)
library(ggplot2)
```

```{r}
KnnSelecting <- function(signals, dists, dist_method = "euclidean", k = 20, noise = 0, index = NULL) {
  stopifnot(noise < k)
  if(is(dists, "dist")) {
    knn1 <- dbscan::kNN(dists, k = k)
    knn2 <- apply(knn1$id, 2, function(x) x %in% which(attr(dists, "Labels") %in% grep("^Noise", rownames(knn1$id), value = TRUE)))
  } else {
    if(is(dists, "kNN")) {
      if(k > dists$k) stop("K too big")
      knn1 <- dists
      knn2 <- apply(knn1$id[, 1:k], 2, function(x) x %in% which(grepl("^Noise", rownames(knn1$id))))
    } else {
      knn1 <- dbscan::kNN(dists[[dist_method]], k = k)
      knn2 <- apply(knn1$id, 2, function(x) x %in% which(attr(dists[[dist_method]], "Labels") %in% grep("^Noise", rownames(knn1$id), value = TRUE)))
    }
  }
  
  tu <- row.names(knn1$id)[which(rowSums(knn2) <= noise)]
  # tu <- union(tu, knn2[which(rowSums(knn2) <= noise), 1])
  tu <- grep("^Noise", tu, invert = TRUE, value = TRUE)
  ggplot(signals, aes(x = Blockade, y = DwellTime, colour = !ID %in% tu)) + 
    geom_point(size = .5) + 
    scale_y_log10() + 
    scale_colour_manual(values = c("#E41A1C", "#377EB8")) + 
    theme_classic() + 
    theme(legend.position = "none") -> p1
  ggplot(signals[ID %in% tu], aes(x = Blockade, y = DwellTime, colour = ID %in% tu)) + 
    geom_point(size = .5) + 
    scale_y_log10() + 
    scale_colour_manual(values = "#E41A1C") + 
    theme_classic() + 
    theme(legend.position = "none") + 
    labs(title = length(tu)) -> p2
  ggplot(signals[!ID %in% tu], aes(x = Blockade, y = DwellTime, colour = ID %in% tu)) + 
    geom_point(size = .5) + 
    scale_y_log10() + 
    scale_colour_manual(values = "#377EB8") + 
    theme_classic() + 
    theme(legend.position = "none") -> p3
  if(!is.null(index)) {
    p1 <- p1 + geom_vline(xintercept = index, colour = "red")
    p2 <- p2 + geom_vline(xintercept = index, colour = "red")
    p3 <- p3 + geom_vline(xintercept = index, colour = "red")
  }
  print(p1 + p2 + p3)
  return(tu)
}

RemoveOutlier <- function(tree, n = 10, Keep = 0.95) {
  h <- length(tree$height)
  g <- cutree(tree, h = tree$height[h])
  g <- data.table(tip = names(g), cluster = g)
  
  while (g[, .N, cluster][N >= n, sum(N)/length(tree$labels)] > Keep) {
    h <- h - 1
    g <- cutree(tree, h = tree$height[h])
    g <- data.table(tip = names(g), cluster = g)
  }
  cat(paste0("Keep: ", round(g[, .N, cluster][N >= n, sum(N)/length(tree$labels)] * 100, 2), "%"))
  g[cluster %in% g[, .N, cluster][N >= n, cluster], tip]
}
```

```{r}
meta <- as.data.table(openxlsx::read.xlsx("data/ChenShanchuan/20231224/20231224 整理数据(20230831).xlsx", sheet = 1))
colnames(meta) <- c("file_name", "start_time", "end_time", "product", "experiment", "sample", "baseline", "note", "X")
setkey(meta, file_name, start_time)
meta <- meta[, .SD[, .(amino_acid = product, start_time, end_time, file_id = paste0(file_name, ".", seq_len(.N)))], file_name]
meta[, sg_files := paste0("./analysis/81.ABFProcessing/RawSignal/RawSignal_", file_name, ".txt")]
meta[, file.exists(sg_files)]
meta <- meta[file.exists(sg_files), ]
meta[, start_time := as.numeric(start_time)]
meta[, end_time := as.numeric(end_time)]
```

# base line distribution

```{r}
RawSig <- lapply(meta[, sg_files], fread)
RawSig <- data.table(file_id = rep(gsub(".txt", "", gsub("RawSignal_", "", basename(meta[, sg_files]))), mapply(nrow, RawSig)), do.call(rbind, RawSig))
```

```{r fig.width=18, fig.height=8}
ggplot(RawSig[BaseMean < 130], aes(x = BaseMean)) + 
  geom_histogram(binwidth = .5) + 
  facet_wrap(~ file_id, scales = "free", nrow = 2)
```

# Raw signals

```{r}
Sigs1 <- lapply(list.files("./analysis/93.Revision/01.StandardAAMix/02.SelectedL0", full.names = TRUE), fread)
Sigs1 <- do.call(rbind, Sigs1)
```

```{r fig.width=18, fig.height=4}
ggplot(Sigs1[Blockade < 0.3], aes(x = Blockade, y = DwellTime)) + 
  geom_point(size = .2) + 
  scale_y_log10() + 
  geom_hline(yintercept = 0.35) + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LNFAE", ""))], Blockade]) + 
  scale_x_continuous(breaks = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LNFAE", ""))], Blockade], labels = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LNFAE", ""))], AA]) + 
  facet_wrap( ~ A, nrow = 1)
```

```{r}
meta0 <- meta[file_id == "2023_12_24_0001.1"]
```

```{r}
meta0[, sig_file := paste0("./analysis/93.Revision/01.StandardAAMix/02.SelectedL0/", file_id, ".MainL0.txt")]
meta0[, file.exists(sig_file)]
meta0 <- meta0[file.exists(sig_file)]
bgfile <- meta0[is.na(amino_acid), sig_file]
bgsignals <- do.call(rbind, lapply(bgfile, fread))
bgsignals[, ID := paste0("Noise_", ID)]
bgsignals <- bgsignals[Blockade < 0.3 & Blockade > 0.1]

B_file <- unique(meta0[is.na(amino_acid), file_name])
B_file <- paste0("./analysis/81.ABFProcessing/FeatureMatrix/FeatureMatrix_", B_file, ".txt")
Big_FeatureMatrix <- do.call(rbind, lapply(B_file, fread))
Big_FeatureMatrix[, ID := paste0("Noise_", ID)]
Big_FeatureMatrix <- Big_FeatureMatrix[ID %in% bgsignals$ID]
```

```{r}
ggplot(bgsignals, aes(x = Blockade, y = DwellTime)) + 
  geom_point(size = .2) + 
  scale_y_log10() + 
  geom_hline(yintercept = 0.35) + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LNFAE", ""))], Blockade]) + 
  scale_x_continuous(breaks = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LNFAE", ""))], Blockade], labels = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LNFAE", ""))], AA]) + 
  facet_wrap( ~ A, nrow = 1)
```

# Select signal from KNN of euclidean distance

```{r}
dist <- readRDS("./analysis/93.Revision/01.StandardAAMix/03.SignalDistance/euclidean_distance_knn.Rds")
```

```{r fig.width=24, fig.height=6}
sig2 <- KnnSelecting(signals = Sigs1, dists = dist, k = 10, noise = 0, index = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("EAFNL", ""))], Blockade])
```

```{r}
dist1 <- readRDS("./analysis/93.Revision/01.StandardAAMix/03.SignalDistance/2023_12_24_0002.1_euclidean_distance_knn.Rds")
```

```{r fig.width=24, fig.height=6}
sig1 <- KnnSelecting(signals = Sigs1[A == "2023_12_24_0002.1"], dists = dist1, k = 15, noise = 1, index = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("EAFNL", ""))], Blockade])
```

```{r}
dist2 <- readRDS("./analysis/93.Revision/01.StandardAAMix/03.SignalDistance/2023_12_24_0003.1_euclidean_distance_knn.Rds")
```

```{r fig.width=24, fig.height=6}
sig2 <- KnnSelecting(signals = Sigs1[A == "2023_12_24_0003.1"], dists = dist2, k = 10, noise = 1, index = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("EAFNL", ""))], Blockade])
```

```{r}
dist3 <- readRDS("./analysis/93.Revision/01.StandardAAMix/03.SignalDistance/2023_12_24_0004.1_euclidean_distance_knn.Rds")
```

```{r fig.width=24, fig.height=6}
sig3 <- KnnSelecting(signals = Sigs1[A == "2023_12_24_0004.1"], dists = dist3, k = 20, noise = 0, index = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("EAFNL", ""))], Blockade])
```

```{r fig.width=12, fig.height=4}
ggplot(Sigs1[ID %in% c(sig1, sig2, sig3)], aes(x = Blockade, y = DwellTime)) + 
  geom_point(size = .2) + 
  scale_y_log10() + 
  geom_hline(yintercept = 0.35) + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LNFAE", ""))], Blockade]) + 
  scale_x_continuous(breaks = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LNFAE", ""))], Blockade], labels = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LNFAE", ""))], AA]) + 
  facet_wrap( ~ A, nrow = 1)
```

# Remove outliers

```{r}
B_file <- unique(meta[!is.na(amino_acid), file_name])
B_file <- paste0("./analysis/81.ABFProcessing/FeatureMatrix/FeatureMatrix_", B_file, ".txt")
B_file <- B_file[file.exists(B_file)]
Sig_FeatureMatrixs <- do.call(rbind, lapply(B_file, fread))
Sig_FeatureMatrixs <- Sig_FeatureMatrixs[ID %in% Sigs1$ID]
FeatureMatrix <- data.frame(Sig_FeatureMatrixs[, grepl("^X", colnames(Sig_FeatureMatrixs)) | colnames(Sig_FeatureMatrixs) %in% c("DwellTime", "DeltaMean", "StageSD", "CurrentSD", "SignalCurrentPercent", "SignalCurrentWidth", "Blockade"), with = F], row.names = Sig_FeatureMatrixs[[1]])
```

```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sig1)
res.ds <- stats::dist(subtab, method = "euclidean")
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sig1s <- RemoveOutlier(tree = res.hc, n = 30, Keep = 0.95)
ggplot(Sigs1[ID %in% sig1s], aes(x = Blockade, y = DwellTime, colour = ID %in% sig1s)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_colour_manual(values = "#E41A1C") + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("EAFNL", ""))], Blockade], colour = "red") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sig1s))
```



```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sig2)
res.ds <- stats::dist(subtab, method = "euclidean")
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sig2s <- RemoveOutlier(tree = res.hc, n = 30, Keep = 0.95)
ggplot(Sigs1[ID %in% sig2s], aes(x = Blockade, y = DwellTime, colour = ID %in% sig2s)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_colour_manual(values = "#E41A1C") + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("EAFNL", ""))], Blockade], colour = "red") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sig1s))
```



```{r}
subtab <- subset.data.frame(FeatureMatrix, row.names(FeatureMatrix) %in% sig3)
res.ds <- stats::dist(subtab, method = "euclidean")
res.hc <- stats::hclust(res.ds, method = "single")
```

```{r}
sig3s <- RemoveOutlier(tree = res.hc, n = 50, Keep = 0.9)
ggplot(Sigs1[ID %in% sig3s], aes(x = Blockade, y = DwellTime, colour = ID %in% sig3s)) + 
  geom_point(size = .5) + 
  scale_y_log10() + 
  scale_colour_manual(values = "#E41A1C") + 
  geom_vline(xintercept = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("EAFNL", ""))], Blockade], colour = "red") + 
  theme_classic() + 
  theme(legend.position = "none") + 
  labs(title = length(sig1s))
```

```{r}
fwrite(Sigs1[ID %in% c(sig1s, sig2s, sig3s)], "./analysis/93.Revision/01.StandardAAMix/04.SelectedSignals/2023_12_24_EAFNL_signals.txt", sep = "\t", quote = FALSE)
```

# Amino acid prediction

```{r}
model1 <- readRDS("./analysis/81.ABFProcessing/SelectedSignals/01.StandardAA/06.Modeling/Version2/L1/SubAAs/ModelAEFLN/RFmodel.Rds")
```

```{r}
selected.signals <- Sigs1[ID %in% c(sig1s, sig2s, sig3s)]
```

```{r}
stopifnot(mean(selected.signals$ID %in% row.names(FeatureMatrix)) == 1)
subtab <- Sig_FeatureMatrixs[ID %in% selected.signals$ID]
pred1 <- data.table(ID = subtab$ID, 
                    Pred1 = as.character(predict(model1, subtab)), 
                    Prob1 = apply(predict(model1, subtab, type = "prob"), 1, max), 
                    Delta1 = apply(predict(model1, subtab, type = "prob"), 1, function(x) - diff(head(sort(x, decreasing = T), 2))))
```

```{r}
EAFNL <- merge(selected.signals, pred1, by = "ID")
```

```{r}
fwrite(EAFNL, "./analysis/93.Revision/01.StandardAAMix/04.SelectedSignals/2023_12_24_EAFNL_signals.txt", sep = "\t", quote = FALSE)
```


```{r fig.width=12, fig.height=4}
ggplot(EAFNL, aes(x = Blockade, y = DwellTime, colour = Pred1)) + 
  geom_point(size = .2) + 
  scale_y_log10() + 
  scale_x_continuous(breaks = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LNFAE", ""))], Blockade], labels = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LNFAE", ""))], AA]) + 
  facet_wrap( ~ A, nrow = 1)
```


```{r fig.width=12, fig.height=4}
ggplot(EAFNL[Delta1 > 0.9], aes(x = Blockade, y = DwellTime, colour = Pred1)) + 
  geom_point(size = .2) + 
  scale_y_log10() + 
  scale_x_continuous(breaks = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LNFAE", ""))], Blockade], labels = AABlockade[AA %in% AMINO_ACID_CODE[unlist(strsplit("LNFAE", ""))], AA]) + 
  facet_wrap( ~ A, nrow = 1)
```

```{r}
EAFNL[Delta1 > 0.9, .N, .(A, Pred1)]
```

# Valid time
